Reorganize parser tests (#63)
* test(parser): Reorganize parser tests

+ Move common parser tests of JVM and JS parser flavor to `ParserTest`
+ Extract flavor-specific values in similar parser tests and move them to `ParserTest`

* test(parser): Reorganize parser tests after merge from `main`

+ Move common parser tests of JVM and JS parser flavor to `ParserTest`

* chore(reformat): Move fields to the top of `ParserTest`
